refactor(media): Refactor how to embed media
