+++ packages/editor/src/plugins/media/services/geogebra/index.tsx
@@ -1,4 +1,4 @@
-import { Embed, Hosting, ResourceResolver } from '../types'
+import { Embed, EmbeddingRenderer, Hosting, ResourceResolver } from '../types'
 
 export const geogebraResourceResolver: ResourceResolver<
   Hosting.GeoGebra,
@@ -8,4 +8,14 @@ export const geogebraResourceResolver: ResourceResolver<
     type: Embed.GeoGebraApplet,
     appletId: resource.appletId,
   }
+}
+
+export const GeogebraRenderer: EmbeddingRenderer<Embed.GeoGebraApplet> = ({
+  appletId,
+}) => {
+  return (
+    <div className="rounded border bg-green-50">
+      Here would be the embedding for GeoGebra applet <code>{appletId}</code>
+    </div>
+  )
 }
